SkywallV0.6: Enemy King Buckets
Enemy King Buckets

Impelemented Enemy King Buckets for the PSTs. Now, they're categorized by friendly king side and enemy king side.

Elo: 18.42 +/- 8.83, nElo: 26.15 +/- 12.51
LOS: 100.00 %, DrawRatio: 44.13 %, PairsRatio: 1.27

Compared to V5:
Score of SkywallV0.6 vs SkywallV0.5: 212 - 41 - 36  [0.796] 289
...      SkywallV0.6 playing White: 134 - 4 - 8  [0.945] 146
...      SkywallV0.6 playing Black: 78 - 37 - 28  [0.643] 143
...      White vs Black: 171 - 82 - 36  [0.654] 289
Elo difference: 236.4 +/- 45.1, LOS: 100.0 %, DrawRatio: 12.5 %
SPRT: llr 3.02 (102.6%), lbound -2.94, ubound 2.94 - H1 was accepted
Games: 2964, Wins: 1109, Losses: 952, Draws: 903, Points: 1560.5 (52.65 %)
Ptnml(0-2): [80, 285, 654, 324, 139]
LLR: 2.90 (-2.25, 2.89) [0.00, 5.00]

+++ Skywall/perft.cpp
@@ -122,7 +122,7 @@ void completePerftTest() {
 
 		//testBoard.precomputeDistances();
 		size_t pairCount = perftTestResults[line].size();
-		for (int z = 0; z < pairCount; z++) {
+		for (size_t z = 0; z < pairCount; z++) {
 			int depth = perftTestResults[line][z].first;
 			uint64_t moveCount = perftTestResults[line][z].second;
